fix(foundry_script): Wildcard erased widths in trait witness matching
Cross-script trait-witness matching decoded a scalar int/uint/long/ulong
parameter's width as wide through type_from_property()'s PropertyInfo
boundary, rejecting a genuinely matching declared int on both sides.

Thread the implementation MethodInfo's argument/return metadata into
type_from_property() so a native or builtin witness recovers its real
declared width when the MethodInfo actually carries it, and reuse
DataType::as_container_slot_type() to turn a carrier-erased width into
an unconstrained wildcard instead of a fabricated one. A genuinely
knowable mismatch (both sides carry real width metadata and disagree)
is still rejected, as caught by Array.size()'s true 32-bit C++ return
type against a trait requirement that mistakenly declared "long".

Closes #1768

+++ modules/foundry_script/tests/scripts/analyzer/features/rtc_builtin_counter.notest.fs
@@ -4,4 +4,4 @@ trait_name RtcBuiltinCounter
 
 abstract func doubled() -> int
 
-abstract func size() -> long
+abstract func size() -> int
